refactor(platform): extract shared draft-attachments resolver
Extract the duplicated draft-to-request payload shaping into a shared
`prepareUserMessageFromDraft$` command and a single `ATTACH_ONLY_PLACEHOLDER`
constant. Both the in-thread send path (`sendMessage$`) and the new-thread
send path (`sendNewThreadMessage$`) now funnel through the same helper, and
the UI's placeholder-strip check uses the same constant.

Addresses PR review round 1 (P1).
